requirements.txt
# --find-links https://download.pytorch.org/whl/torch_stable.html 
torchvision
torchaudio
# torch==1.10.0+cu113
albumentations
effdet
timm
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
antlr4-python3-runtime==4.8
arrow==1.2.1
astunparse==1.6.3
async-timeout==4.0.0
binaryornot==0.4.4
cachetools==4.2.4
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
coco-froc-analysis==0.0.29
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.1.0
cookiecutter==1.7.3
cython==0.29.24
distro==1.6.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
ensemble-boxes==1.0.7
evalutils==0.3.0
flatbuffers==2.0
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
gitpython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.6.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3
imageio==2.10.3
install==1.3.4
jinja2-time==0.2.0
joblib==1.1.0
keras==2.7.0
keras-preprocessing==1.1.2
libclang==12.0.0
llvmlite==0.37.0
markdown==3.3.4
monai==0.7.0
multidict==5.2.0
names==0.3.0
networkx==2.6.3
nibabel==3.2.1
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.1
opencxr==1.2.0
opt-einsum==3.3.0
packaging==21.2
pathtools==0.1.2
plotly==5.4.0
poyo==0.5.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydeprecate==0.3.1
pydicom==2.2.2
pyerf==1.0.1
pyparsing==2.4.7
pypng==0.0.21
python-dotenv==0.19.1
python-slugify==5.0.2
pytorch-lightning==1.5.1
pytorch-ranger==0.1.1
pywavelets==1.2.0
pyyaml==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.13.0
rsa==4.7.2
scikit-build==0.12.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
sentry-sdk==1.4.3
shortuuid==1.0.7
smmap==5.0.0
subprocess32==3.5.4
tenacity==8.0.1
# tensorboard==2.7.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.0
# tensorflow==2.7.0
# # tensorflow-estimator==2.7.0
# tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
tifffile==2021.11.2
torch-optimizer==0.3.0
torchmetrics==0.6.0
torchxrayvision==0.0.32
tqdm==4.62.3
tsnecuda==0.1.1
urllib3==1.26.7
wandb==0.12.7
werkzeug==2.0.2
wget==3.2
wrapt==1.13.3
yarl==1.7.2
yaspin==2.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html 